Rename Twisted's HTTP CONNECT proxy support to distinguish
from asyncio.

Little bit 'sad' having to have two different kinds of HTTP
CONNECT proxy support in one repo, but that's how it is for now.
